chore: Clean up project contributor edit page
