+++ assets/js/pages/ProjectContributorsPage/index.tsx
@@ -2,7 +2,6 @@ import React from "react";
 
 import * as Paper from "@/components/PaperContainer";
 import * as Pages from "@/components/Pages";
-import * as Icons from "@tabler/icons-react";
 import * as Projects from "@/models/projects";
 
 import { PrimaryButton, SecondaryButton } from "@/components/Buttons";
